Add files via upload
Slotted Aloha Simulation:
Aloha is a congestion based networking MAC protocol. In case of ALOHA, each packet is of equal length. ALOHA is further divided into:
Pure Aloha: Each packet is allowed to be sent at any interval. So, comparatively, there is more probability of collision and because of that, its throughput is almost 18%. 
Slotted Aloha: Each packet is allowed to be only at the end of a particular slot. So, comparatively, there is lesser probability of collision and because of that, its throughput is almost 36%.
I have created a simulation of slotted ALOHA, created different graphs to show efficiency of SLOTTED ALOHA and then have compared it with Pure Aloha. 
The simulation is created as a graphical user interface GUIDE been provided in MATLAB 
So, the simulation contains following:
- GUIDE  for Graphical User Interface
- 2D Plots
- Creation of smaller filled circles as nodes on Interface
- Comparision of Pure and Slotted ALOHA